✨ feature : mypage css/view/component 분리리
