+++ frontend/linki-user/src/assets/css/mypage-sidebar.css
@@ -1,0 +1,70 @@
+.mypage-sidebar {
+  width: 260px;
+  padding: 48px 0 0 60px;
+  background: #fff;
+  font-family: 'Pretendard', 'Noto Sans KR', sans-serif;
+  border-right: 1px solid #eee;
+}
+
+.mypage-sidebar .sidebar-title {
+  color: #bdbdbd;
+  font-size: 16px;
+  font-weight: 500;
+  margin-bottom: 40px;
+}
+
+.mypage-sidebar .sidebar-title .divider {
+  margin: 0 8px;
+}
+
+.mypage-sidebar .sidebar-title .mypage {
+  color: #222;
+  font-weight: 400;
+}
+
+.mypage-sidebar .sidebar-nav {
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+}
+
+.mypage-sidebar .sidebar-section {
+  color: #222;
+  font-size: 16px;
+  font-weight: 600;
+  margin-bottom: 8px;
+}
+
+.mypage-sidebar .sidebar-section ul {
+  margin: 8px 0 0 0;
+  padding: 0;
+  list-style: none;
+}
+
+.mypage-sidebar .sidebar-section li {
+  color: #bdbdbd;
+  font-size: 15px;
+  font-weight: 400;
+  margin-bottom: 6px;
+  cursor: pointer;
+  transition: color 0.2s;
+  padding-left: 12px;
+}
+
+.mypage-sidebar .sidebar-section li.active,
+.mypage-sidebar .sidebar-section li:hover {
+  color: #9B51E0;
+  font-weight: 600;
+}
+
+.mypage-sidebar .sidebar-section li.active > a,
+.mypage-sidebar .sidebar-section li > a {
+  color: inherit;
+  text-decoration: none;
+  transition: color 0.2s;
+}
+
+/* 추가된 호버 효과 */
+.mypage-sidebar .sidebar-section li:hover > a {
+  color: #9B51E0;
+}
