+++ frontend/linki-user/src/components/user/advertiser/mypage/MyPageProfile.vue
@@ -1,0 +1,30 @@
+<template>
+  <main class="profile-content">
+    <h1 class="profile-title">프로필 / 비밀번호</h1>
+    <div class="profile-box">
+      <!-- 프로필 설정 영역 -->
+      <section class="profile-section">
+        <h2 class="section-title">프로필 설정</h2>
+        <div class="profile-form">
+          <!-- 프로필 폼 내용 추가 예정 -->
+        </div>
+      </section>
+
+      <!-- 비밀번호 변경 영역 -->
+      <section class="profile-section">
+        <h2 class="section-title">비밀번호 변경</h2>
+        <div class="password-form">
+          <!-- 비밀번호 변경 폼 내용 추가 예정 -->
+        </div>
+      </section>
+    </div>
+  </main>
+</template>
+
+<script setup>
+// 프로필 관련 로직 추가 예정
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
