+++ frontend/linki-user/src/components/user/advertiser/mypage/MyPageSubscription.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="subscription-content">
+    <h1 class="content-title">구독 상태 확인</h1>
+    <div class="content-box">
+      <!-- 구독 상태 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageSubscription'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
