+++ frontend/linki-user/src/components/user/influencer/mypage/MyPageBasicInfo.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="basic-info-content">
+    <h1 class="content-title">기본 정보</h1>
+    <div class="content-box">
+      <!-- 기본 정보 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageBasicInfo'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
