+++ frontend/linki-user/src/components/user/influencer/mypage/MyPageChannelInfo.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="channel-info-content">
+    <h1 class="content-title">채널 등록</h1>
+    <div class="content-box">
+      <!-- 채널 등록 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageChannelInfo'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
