+++ frontend/linki-user/src/components/user/influencer/mypage/MyPageCompletedContracts.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="completed-contracts-content">
+    <h1 class="content-title">완료된 계약</h1>
+    <div class="content-box">
+      <!-- 완료된 계약 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageCompletedContracts'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
