+++ frontend/linki-user/src/components/user/influencer/mypage/MyPageOngoingContracts.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="ongoing-contracts-content">
+    <h1 class="content-title">진행중인 계약</h1>
+    <div class="content-box">
+      <!-- 진행중인 계약 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageOngoingContracts'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
