+++ frontend/linki-user/src/components/user/influencer/mypage/MyPageReceivedReviews.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="received-reviews-content">
+    <h1 class="content-title">받은 리뷰</h1>
+    <div class="content-box">
+      <!-- 받은 리뷰 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageReceivedReviews'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
