+++ frontend/linki-user/src/components/user/influencer/mypage/MyPageSettlement.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="settlement-content">
+    <h1 class="content-title">정산 내역</h1>
+    <div class="content-box">
+      <!-- 정산 내역 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageSettlement'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
