+++ frontend/linki-user/src/components/user/influencer/mypage/MyPageSubscriptionApply.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="subscription-apply-content">
+    <h1 class="content-title">구독 신청</h1>
+    <div class="content-box">
+      <!-- 구독 신청 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageSubscriptionApply'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
