+++ frontend/linki-user/src/components/user/influencer/mypage/MyPageSubscriptionManage.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="subscription-manage-content">
+    <h1 class="content-title">나의 구독 관리</h1>
+    <div class="content-box">
+      <!-- 구독 관리 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageSubscriptionManage'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
