+++ frontend/linki-user/src/components/user/influencer/mypage/MyPageWrittenReviews.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="written-reviews-content">
+    <h1 class="content-title">작성한 리뷰</h1>
+    <div class="content-box">
+      <!-- 작성한 리뷰 내용 추가 예정 -->
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'MyPageWrittenReviews'
+}
+</script>
+
+<style>
+@import '@/css/mypage.css';
+</style>
