+++ frontend/linki-user/src/views/user/common/MyPage.vue
@@ -1,4 +1,4 @@
-<script setup>
+<!-- <script setup>
 import { useUserStore } from '@/stores/user'
 import { onMounted } from 'vue'
 import { useRouter } from 'vue-router'
@@ -24,4 +24,4 @@ onMounted(() => {
 
 <style scoped>
 
-</style>
+</style> -->
